+++ App.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import HomePage from "./src/pages/HomePage";
-import Products from "./src/pages/Products";
+import Product from "./src/pages/Product";
 import Pricing from "./src/pages/Pricing";
 import Header from "./src/components/navigations/header/Header";
 import Footer from "./src/components/navigations/footer/Footer";
@@ -16,7 +16,7 @@ function App() {
         <main >
           <Routes>
             <Route path="/" element={<HomePage />} />
-            <Route path="products" element={<Products />} />
+            <Route path="products" element={<Product />} />
             <Route path="pricing" element={<Pricing />} />
             <Route path="app" element={<AppLayout />} />
             <Route path="*" element={<PageNotFound />} />
