+++ src/index.css
@@ -65,7 +65,7 @@ input:focus {
   font-size: 1.6rem;
   font-weight: 600;
   padding: 1rem 3rem;
-  border-radius: 5px;
+  border-radius: 10px;
 }
 
 /*
