+++ src/pages/AppLayout.module.css
@@ -1,0 +1,7 @@
+.app {
+  height: 100vh;
+  padding: 2.4rem;
+  overscroll-behavior-y: none;
+  display: flex;
+  position: relative;
+}
