+++ src/pages/HomePage.jsx
@@ -1,14 +1,22 @@
-import React from 'react'
-import { Link } from 'react-router-dom';
+import { Link } from "react-router-dom";
+import styles from "./HomePage.module.css";
 
-const HomePage = () => {
+export default function Homepage() {
   return (
-    <div>
-      <p>Home Page</p>
-
-      <Link to="/app"> go to app layout</Link>
-    </div>
-  )
+    <main className={styles.homepage}>
+      <section>
+        <h1>
+          You travel the world.
+          <br />
+          WorldWise keeps track of your adventures.
+        </h1>
+        <h2>
+          A world map that tracks your footsteps into every city you can think
+          of. Never forget your wonderful experiences, and show your friends how
+          you have wandered the world.
+        </h2>
+        <Link to="/app" className="cta">Start Tracking Now</Link>
+      </section>
+    </main>
+  );
 }
-
-export default HomePage
