+++ src/pages/Pricing.jsx
@@ -1,11 +1,24 @@
-import React from 'react'
+// Uses the same styles as Product
+import styles from "./Product.module.css";
 
-const Pricing = () => {
+export default function Product() {
   return (
-    <div>
-     Pricing 
-    </div>
-  )
+    <main className={styles.product}>
+      <section>
+        <div>
+          <h2>
+            Simple pricing.
+            <br />
+            Just $9/month.
+          </h2>
+          <p>
+            Lorem ipsum dolor, sit amet consectetur adipisicing elit. Vitae vel
+            labore mollitia iusto. Recusandae quos provident, laboriosam fugit
+            voluptatem iste.
+          </p>
+        </div>
+        <img src="img-2.jpg" alt="overview of a large city with skyscrapers" />
+      </section>
+    </main>
+  );
 }
-
-export default Pricing
