--- src/pages/Products.jsx
@@ -1,12 +1,0 @@
-import React from 'react'
-
-
-const Products = () => {
-  return (
-    <div>
-      products
-    </div>
-  )
-}
-
-export default Products
